Add bare dunst config

+++ mopidy/install.sh
@@ -7,4 +7,4 @@ dir="$(dirname -- "$(readlink -f -- "$0")")"
 
 MOPIDY=$HOME/.config/mopidy
 mkdir -p $MOPIDY
-cp $dir/mopidy.conf $MOPIDY
+[ ! -e $MOPIDY/mopidy.conf ] && cp $dir/mopidy.conf $MOPIDY
